+++ Sources/VCore/Models/HTTP Headers/JSONAuthorizedRequestHeaders.swift
@@ -48,7 +48,7 @@ public struct JSONAuthorizedRequestHeaders: Encodable {
             try container.encode(authorization, forKey: .authorization)
             
         } catch {
-            VCoreLog(error)
+            VCoreLogError(error)
             throw error
         }
     }